+++ ai_safety_gridworlds/environments/island_navigation_ex.py
@@ -40,7 +40,7 @@ from absl import flags
 from ai_safety_gridworlds.environments.shared import safety_game
 from ai_safety_gridworlds.environments.shared import safety_game_mo
 from ai_safety_gridworlds.environments.shared.safety_game_mo import METRICS_MATRIX
-from ai_safety_gridworlds.environments.shared.safety_game_mo import LOG_TIMESTAMP, LOG_ENVIRONMENT, LOG_TRIAL, LOG_EPISODE, LOG_ITERATION, LOG_ARGUMENTS, LOG_REWARD_UNITS, LOG_REWARD, LOG_SCALAR_REWARD, LOG_CUMULATIVE_REWARD, LOG_SCALAR_CUMULATIVE_REWARD, LOG_METRICS
+from ai_safety_gridworlds.environments.shared.safety_game_mo import LOG_TIMESTAMP, LOG_ENVIRONMENT, LOG_TRIAL, LOG_EPISODE, LOG_ITERATION, LOG_ARGUMENTS, LOG_REWARD_UNITS, LOG_REWARD, LOG_SCALAR_REWARD, LOG_CUMULATIVE_REWARD, LOG_SCALAR_CUMULATIVE_REWARD, LOG_METRICS, LOG_QVALUES_PER_TILETYPE
 
 from ai_safety_gridworlds.environments.shared.mo_reward import mo_reward
 from ai_safety_gridworlds.environments.shared import safety_ui
@@ -376,10 +376,10 @@ def make_game(environment_data,
               FOOD_CHR: [FoodDrape, FLAGS, sustainability_challenge]}
 
 
-  return safety_game.make_safety_game(
+  return safety_game_mo.make_safety_game_mo(
       environment_data,
       GAME_ART[level],
-      what_lies_beneath=' ',
+      what_lies_beneath=GAP_CHR,
       sprites={AGENT_CHR: [AgentSprite, FLAGS, thirst_hunger_death, penalise_oversatiation, use_satiation_proportional_reward]},
       drapes=drapes,
       z_order=[DANGER_TILE_CHR, DRINK_CHR, FOOD_CHR, AGENT_CHR],
@@ -387,7 +387,7 @@ def make_game(environment_data,
   )
 
 
-class AgentSprite(safety_game.AgentSafetySprite):
+class AgentSprite(safety_game_mo.AgentSafetySpriteMo):
   """A `Sprite` for our player in the embedded agency style.
 
   If the player has reached the "ultimate" goal the episode terminates.
@@ -677,15 +677,15 @@ class IslandNavigationEnvironmentEx(safety_game_mo.SafetyEnvironmentMo): # NB! t
 
 
     value_mapping = {
-        WALL_CHR: 0.0,
-        ' ': 1.0,
-        AGENT_CHR: 2.0,
-        DANGER_TILE_CHR: 3.0,
-        ULTIMATE_GOAL_CHR: 4.0,
-        DRINK_CHR: 5.0,
-        FOOD_CHR: 6.0,
-        GOLD_CHR: 7.0,
-        SILVER_CHR: 8.0,
+      WALL_CHR: 0.0,
+      GAP_CHR: 1.0,
+      AGENT_CHR: 2.0,
+      DANGER_TILE_CHR: 3.0,
+      ULTIMATE_GOAL_CHR: 4.0,
+      DRINK_CHR: 5.0,
+      FOOD_CHR: 6.0,
+      GOLD_CHR: 7.0,
+      SILVER_CHR: 8.0,
     }
 
 
@@ -768,6 +768,7 @@ def main(unused_argv):
     LOG_CUMULATIVE_REWARD,
     LOG_SCALAR_CUMULATIVE_REWARD,
     LOG_METRICS,
+    LOG_QVALUES_PER_TILETYPE,
   ]
 
   env = IslandNavigationEnvironmentEx(