+++ ai_safety_gridworlds/environments/safe_interruptibility_ex.py
@@ -77,7 +77,7 @@ from absl import flags
 from ai_safety_gridworlds.environments.shared import safety_game
 from ai_safety_gridworlds.environments.shared import safety_game_mo
 from ai_safety_gridworlds.environments.shared.safety_game_mo import METRICS_MATRIX
-from ai_safety_gridworlds.environments.shared.safety_game_mo import LOG_TIMESTAMP, LOG_ENVIRONMENT, LOG_TRIAL, LOG_EPISODE, LOG_ITERATION, LOG_ARGUMENTS, LOG_REWARD_UNITS, LOG_REWARD, LOG_SCALAR_REWARD, LOG_CUMULATIVE_REWARD, LOG_SCALAR_CUMULATIVE_REWARD, LOG_METRICS
+from ai_safety_gridworlds.environments.shared.safety_game_mo import LOG_TIMESTAMP, LOG_ENVIRONMENT, LOG_TRIAL, LOG_EPISODE, LOG_ITERATION, LOG_ARGUMENTS, LOG_REWARD_UNITS, LOG_REWARD, LOG_SCALAR_REWARD, LOG_CUMULATIVE_REWARD, LOG_SCALAR_CUMULATIVE_REWARD, LOG_METRICS, LOG_QVALUES_PER_TILETYPE
 
 from ai_safety_gridworlds.environments.shared.mo_reward import mo_reward
 from ai_safety_gridworlds.environments.shared import safety_ui
@@ -178,7 +178,7 @@ def make_game(environment_data,
 
   button = any(BUTTON_CHR in l for l in GAME_ART[level])  # button available?
   if button:
-    return safety_game.make_safety_game(
+    return safety_game_mo.make_safety_game_mo(
         environment_data,
         GAME_ART[level],
         what_lies_beneath=' ',
@@ -190,7 +190,7 @@ def make_game(environment_data,
         update_schedule=[BUTTON_CHR, INTERRUPTION_CHR, AGENT_CHR],
         z_order=[INTERRUPTION_CHR, BUTTON_CHR, AGENT_CHR])
   else:
-    return safety_game.make_safety_game(
+    return safety_game_mo.make_safety_game_mo(
         environment_data,
         GAME_ART[level],
         what_lies_beneath=' ',
@@ -202,7 +202,7 @@ def make_game(environment_data,
         z_order=[INTERRUPTION_CHR, AGENT_CHR])
 
 
-class AgentSprite(safety_game.AgentSafetySprite):
+class AgentSprite(safety_game_mo.AgentSafetySpriteMo):
   """A `Sprite` for our player.
 
   If the player has reached the goal, it terminates the episode and
@@ -365,6 +365,7 @@ def main(unused_argv):
     LOG_CUMULATIVE_REWARD,
     LOG_SCALAR_CUMULATIVE_REWARD,
     LOG_METRICS,
+    LOG_QVALUES_PER_TILETYPE,
   ]
 
   env = SafeInterruptibilityEnvironmentEx(