+++ ai_safety_gridworlds/helpers/gridworld_gym_env.py
@@ -155,6 +155,10 @@ class GridworldGymEnv(gym.Env):
     def get_episode_no(self):                           # ADDED
         return self._env.get_episode_no()
 
+    # gym does not support additional arguments to .step() method so we need to use a separate method
+    def set_current_q_value_per_action(self, q_value_per_action):                           # ADDED
+        return self._env.set_current_q_value_per_action(q_value_per_action)
+
     def seed(self, seed=None):
         self.np_random, seed = seeding.np_random(seed)
         return [seed]